requirements.txt
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
click==8.1.7
colorama==0.4.6
coverage==7.2.7
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==0.21.1
SQLAlchemy==2.0.21
typing_extensions==4.7.1
Werkzeug==2.2.3
wincertstore==0.2
zipp==3.15.0
